+++ protocol/lombok.config
@@ -1,2 +1,3 @@
 # This file is generated by the 'io.freefair.lombok' Gradle plugin
 config.stopBubbling = true
+lombok.addLombokGeneratedAnnotation = true
