+++ protocol/src/main/java/com/amazon/opendistroforelasticsearch/sql/protocol/response/ResponseListener.java
@@ -14,7 +14,7 @@
  *
  */
 
-package com.amazon.opendistroforelasticsearch.sql.common.protocol.response;
+package com.amazon.opendistroforelasticsearch.sql.protocol.response;
 
 /**
  * Response listener for response post-processing callback.
